adding md5sum_renaming.py readme and two folders of images as examples to use
