+++ pipeline/naming_files/Pipfile
@@ -1,0 +1,12 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+hashlib = "*"
+
+[requires]
+python_version = "3.7"
