+++ pipeline/naming_files/README.md
@@ -1,0 +1,8 @@
+# Naming image files
+
+## finding md5sum for each image file
+
+## checking for duplicates
+
+## renaming image file by its respective hash + extension using *md5sum* and removing duplicates
+
